requirements.txt
Django==3.1.2
djangorestframework>=3.9
psycopg2-binary>=2.8.6
Pillow